Docs : move styled component
